[stdlib & genesis] refactor stdlib build and extract rooch-genesis crate
