+++ crates/rooch-executor/Cargo.toml
@@ -40,3 +40,4 @@ moveos-store = { workspace = true }
 moveos-types = { workspace = true }
 
 rooch-types = { workspace = true }
+rooch-genesis = { workspace = true }
