+++ crates/rooch-framework-tests/tests/cases/event/basic.move
