+++ crates/rooch-framework-tests/tests/cases/table/basic.exp
