+++ crates/rooch-framework/Move.toml
@@ -10,4 +10,4 @@ vm_reserved = "0x0"
 rooch_association = "0xA550C18"
 
 [dependencies]
-MoveosStdLib = { local = "../moveos-stdlib" }
+MoveosStdLib = { local = "../../moveos/moveos-stdlib/moveos-stdlib" }
