+++ crates/rooch-framework/src/lib.rs
@@ -1,5 +1,4 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod moveos;
-pub mod vm;
+pub mod natives;
