+++ crates/rooch-framework/src/natives/rooch_framework/crypto/hash.rs
