+++ crates/rooch-server/Cargo.toml
@@ -40,7 +40,6 @@ move-core-types = { workspace = true }
 move-resource-viewer = { workspace = true }
 move-binary-format = { workspace = true }
 
-moveos = { workspace = true }
 moveos-store = { workspace = true }
 moveos-types = { workspace = true }
 move-bytecode-utils = { workspace = true }
