+++ crates/rooch/src/commands/move_cli/commands/unit_test.rs
@@ -6,11 +6,10 @@ use move_cli::base::test;
 use move_package::BuildConfig;
 use move_unit_test::extensions::set_extension_hook;
 use move_vm_runtime::native_extensions::NativeContextExtensions;
-use moveos_stdlib::natives::{
-    all_natives, moveos_stdlib::raw_table::NativeTableContext, GasParameters,
-};
+use moveos_stdlib::natives::moveos_stdlib::raw_table::NativeTableContext;
 use moveos_store::state_store::StateDB;
 use once_cell::sync::Lazy;
+use rooch_framework::natives::{all_natives, GasParameters};
 use std::path::PathBuf;
 
 #[derive(Parser)]
