+++ moveos/moveos/src/vm/mod.rs
@@ -3,7 +3,7 @@
 
 //TODO remove this after refactor the caller crates.
 pub mod dependency_order {
-    pub use moveos_stdlib::dependency_order::*;
+    pub use moveos_stdlib_builder::dependency_order::*;
 }
 pub mod moveos_vm;
 pub mod tx_argument_resolver;
